fixed rendering issues by setting priority of player, created better structure for WorldManagers. minor code cleanup.
